ENG-7251 - Refactor RN Callbacks folder to deduplicate code
