+++ NeuroID/src/android/java/com/neuroid/tracker/callbacks/NIDActivityCallbackUtil.kt
@@ -1,0 +1,6 @@
+package com.neuroid.tracker.callbacks
+
+// Native Android needs to register the activities
+internal fun registrationHelpers(r:Runnable){
+    r.run()
+}
