+++ NeuroID/src/reactNative/java/com/neuroid/tracker/callbacks/NIDActivityCallbackUtil.kt
@@ -1,0 +1,6 @@
+package com.neuroid.tracker.callbacks
+
+// In RN users call registerPageTargets so we don't
+//    want to run this code and double register
+internal fun registrationHelpers(r:Runnable){
+}
